Add/check in bounds (#1)
* clarify that bbox class is for tiles

* add optional check if point is inside human roi

This commit adds an optional check if a point is inside another polygon.
Namely, the polygon would be the union of all human-made ROIs of a
certain name. These ROIs are typically placed before annotation begins,
and nuclei are dotted inside these ROIs. The ROIs are typically
rectangular, and this commit assumes they are rectangular.

With this commit, some confusion might arise between bbox and human ROI.
Bbox is the bounding box of the tile (remember the slide is separated
into tiles). The human ROI is a different concept, and is only related
to annotation. A previous commit renamed bbox to something mentioning
tiles.

* add use example + example of movie creation